% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustedSharpeError.R
\name{AdjustedSharpeError}
\alias{AdjustedSharpeError}
\title{Calculates error in the Adjusted Sharpe Ratio
The error calculation is sourced from : http://edge-fund.com/Lo02.pdf and is defined as sqrt(1+((SR)^2)/2) divided by N for the RAW Sharpe
For Adjusted Sharpe, we use the error for Kurtosis and Skewness in conjunction with the Sharpe error}
\usage{
AdjustedSharpeError(R, Rf = 0, ...)
}
\arguments{
\item{Ra, }{Rb ( default zero )
No wrapper used, instead the AdjustedSharpeRatio code is re-used to calculate Sharpe in the function itself
Possible modification : Pass a flag variable error ( default zero ) into all sharpe Ratio functions and return error instead of Sharpe if error = 1}
}
\description{
Calculates error in the Adjusted Sharpe Ratio
The error calculation is sourced from : http://edge-fund.com/Lo02.pdf and is defined as sqrt(1+((SR)^2)/2) divided by N for the RAW Sharpe
For Adjusted Sharpe, we use the error for Kurtosis and Skewness in conjunction with the Sharpe error
}

