% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxDrawdown.R
\name{AverageDrawdown}
\alias{AverageDrawdown}
\title{Calculates the average depth of the observed drawdowns.}
\usage{
AverageDrawdown(R, ...)
}
\arguments{
\item{R}{an xts, vector, matrix, data frame, timeSeries or zoo object of
asset returns}

\item{\dots}{any other passthru parameters}
}
\description{
ADD = abs(sum[j=1,2,...,d](D_j/d)) where
D'_j = jth drawdown over entire period
d = total number of drawdowns in the entire period
}
\author{
Peter Carl
}

