% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAPM.coefficients.R
\name{CAPM.coefficients}
\alias{CAPM.coefficients}
\alias{SFM.coefficients}
\title{calculate single factor model (CAPM) alpha and beta coefficients}
\usage{
CAPM.coefficients(xRa, xRb, subset, ..., method = "LS", family = "mOpt")
}
\arguments{
\item{xRa}{an xts, vector, matrix, data frame, timeSeries or zoo object of
excess asset returns}

\item{xRb}{excess return vector of the benchmark asset}
}
\description{
The single factor model or CAPM is the (alpha,beta) of an asset to the variance 
and covariance of an initial portfolio.  Used to determine diversification potential.
}
\details{
This function uses a linear intercept model to achieve the same results as
the symbolic model used by \code{\link{BetaCoVariance}}. This function is designed to
be used a a wrapper for SFM's regression models. Using this, one can easily fit different
types of linear models (Ordinary Least Squares as well as Robust Linear Regression).

\deqn{\beta_{a,b}=\frac{CoV_{a,b}}{\sigma_{b}}=\frac{\sum((R_{a}-\bar{R_{a}})(R_{b}-\bar{R_{b}}))}{\sum(R_{b}-\bar{R_{b}})^{2}}}{beta
= cov(Ra,Rb)/var(R)}

Ruppert(2004) reports that this equation will give the estimated slope of
the linear regression of \eqn{R_{a}}{Ra} on \eqn{R_{b}}{Rb} and that this
slope can be used to determine the risk premium or excess expected return
(see Eq. 7.9 and 7.10, p. 230-231).
}
\examples{

data(managers)
    CAPM.coefficients(managers[,1,drop=FALSE], 
			managers[,8,drop=FALSE]) 
    CAPM.alpha(managers[,1,drop=FALSE], 
			managers[,8,drop=FALSE], 
			Rf = managers[,10,drop=FALSE])
    CAPM.alpha(managers[,1:6], 
			managers[,8,drop=FALSE], 
			Rf=.035/12)
    CAPM.alpha(managers[,1:6], 
			managers[,8,drop=FALSE], 
			Rf = managers[,10,drop=FALSE])
    CAPM.alpha(managers[,1:6], 
			managers[,8:7,drop=FALSE], 
			Rf=.035/12) 
    CAPM.alpha(managers[,1:6], 
			managers[,8:7,drop=FALSE], 
			Rf = managers[,10,drop=FALSE])
    CAPM.beta(managers[, "HAM2", drop=FALSE], 
			managers[, "SP500 TR", drop=FALSE], 
			Rf = managers[, "US 3m TR", drop=FALSE])
    CAPM.beta.bull(managers[, "HAM2", drop=FALSE], 
			managers[, "SP500 TR", drop=FALSE], 
			Rf = managers[, "US 3m TR", drop=FALSE])
    CAPM.beta.bear(managers[, "HAM2", drop=FALSE], 
			managers[, "SP500 TR", drop=FALSE], 
			Rf = managers[, "US 3m TR", drop=FALSE])
    TimingRatio(managers[, "HAM2", drop=FALSE], 
			managers[, "SP500 TR", drop=FALSE], 
			Rf = managers[, "US 3m TR", drop=FALSE])
    chart.Regression(managers[, "HAM2", drop=FALSE], 
			managers[, "SP500 TR", drop=FALSE], 
			Rf = managers[, "US 3m TR", drop=FALSE], 
			fit="conditional", 
			main="Conditional Beta")
  		
}
\references{
Sharpe, W.F. Capital Asset Prices: A theory of market
equilibrium under conditions of risk. \emph{Journal of finance}, vol 19,
1964, 425-442. \cr Ruppert, David. \emph{Statistics and Finance, an
Introduction}. Springer. 2004. \cr Bacon, Carl. \emph{Practical portfolio
performance measurement and attribution}. Wiley. 2004. \cr
}
\seealso{
\code{\link{BetaCoVariance}} \code{\link{CAPM.alpha}}
\code{\link{CAPM.utils}} \code{\link{CAPM.beta}}
}
\author{
Peter Carl
}
