% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxDrawdown.R
\name{CDD}
\alias{CDD}
\alias{CDaR}
\title{Calculate Uryasev's proposed Conditional Drawdown at Risk (CDD or CDaR)
measure}
\usage{
CDD(
  R,
  weights = NULL,
  geometric = TRUE,
  invert = TRUE,
  p = 0.95,
  method = c("discrete", "average", "quantile"),
  ...
)
}
\arguments{
\item{R}{an xts, vector, matrix, data frame, timeSeries or zoo object of
asset returns}

\item{weights}{portfolio weighting vector, default NULL, see Details}

\item{geometric}{utilize geometric chaining (TRUE) or simple/arithmetic chaining (FALSE) to aggregate returns,
default TRUE}

\item{invert}{TRUE/FALSE whether to invert the drawdown measure.  see
Details.}

\item{p}{confidence level for calculation, default p=0.95}

\item{method}{one of "discrete", "average", or "quantile". See Details.}

\item{\dots}{any other passthru parameters}
}
\description{
For some confidence level \eqn{p}, the conditional drawdown is the mean
of the worst \eqn{p\%} drawdowns.
}
\details{
The \code{method} parameter allows for three mathematical variations of CDD/CDaR:
\itemize{
  \item{\code{discrete}}{ (Default) Computes the Expected Shortfall of the discrete, isolated peak-to-trough drawdown sequences. This isolates major discrete events without penalizing the prolonged duration of shallow drawdowns.}
  \item{\code{average}}{ Computes the mean of the continuous sample path of drawdowns that exceed the threshold. This strictly adheres to the Chekhlov (2003) continuous definition but can be heavily autocorrelated, mathematically overweighting long, shallow drawdowns over sharp, brief crashes.}
  \item{\code{quantile}}{ (Deprecated) A legacy implementation that merely computed the VaR (quantile) of the discrete drawdowns, rather than the Expected Shortfall (mean).}
}
}
\examples{

data(edhec)
t(round(CDD(edhec), 4))

}
\references{
Chekhlov, A., Uryasev, S., and M. Zabarankin. Portfolio
Optimization With Drawdown Constraints. B. Scherer (Ed.) Asset and Liability
Management Tools, Risk Books, London, 2003
https://www.ise.ufl.edu/uryasev/drawdown.pdf
}
\seealso{
\code{\link{ES}} \code{\link{maxDrawdown}}
}
\author{
Brian G. Peterson
}
