\name{chart.RegressionDiagnostics}
\alias{chart.RegressionDiagnostics}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ regression diagnostics charts }
\description{
}
\usage{
chart.RegressionDiagnostics(R, Rb)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{R}{ a vector, matrix, data frame, timeSeries or zoo object of asset returns }
  \item{Rb}{ return vector of the benchmark asset }
}
\details{
}
\value{
chart of regression stats to help you validate the regression outputs
}
\references{ }
\author{ Peter Carl }
\seealso{
    \code{\link{charts.RollingRegression}} \cr
    \code{\link{chart.RollingRegression}} \cr
    \code{\link{chart.Histogram}} \cr
    \code{\link{chart.QQPlot}}    \cr
    \code{\link{chart.Boxplot}}
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ts }
\keyword{ multivariate }
\keyword{ distribution }
\keyword{ models }
\keyword{ hplot }
