% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.SFM.R
\name{chart.SFM}
\alias{chart.SFM}
\alias{chart.CAPM}
\title{Compare CAPM estimated using robust estimators with that estimated by OLS}
\usage{
chart.SFM(
  Ra,
  Rb,
  Rf = 0,
  fit.models.chart = F,
  which.plots = NULL,
  main = "lm vs lmRobdetMM",
  ylim = NULL,
  xlim = NULL,
  family = c("mopt", "opt", "bisquare"),
  legend.loc = "topleft",
  makePct = FALSE,
  lm.outliers = F
)
}
\arguments{
\item{Ra}{an xts, vector, matrix, data frame, timeSeries or zoo object of
asset returns}

\item{Rb}{return vector of the benchmark asset}

\item{Rf}{risk free rate, in same period as your returns}

\item{fit.models.chart}{option to output charts using fit.models package. Defaults to False}

\item{which.plots}{If fit.models.chart is set to TRUE, then this is a list of 
the plots that the user wants to see}

\item{main}{Title of the generated plot. Defaults to "lm vs lmRobdetMM"}

\item{ylim}{Limits on the y-axis of the plots. Defaults to min-max}

\item{xlim}{Limits on the x-axis of the plots. Defaults to min-max}

\item{family}{(Optional): 
This is a string specifying the name of the family of loss function
to be used (current valid options are "bisquare", "opt" and "mopt").
Incomplete entries will be matched to the current valid options. 
Defaults to "mopt".}

\item{legend.loc}{Position of legends. See plot() function for more info.}

\item{makePct}{If Returns should be converted to percentage. Defaults to False}

\item{lm.outliers}{If outlier boundaries are to be shown with respect to lm model,
then set this to true. Defaults to false and outlier boundaries are shown wrt
lmrobdetMM model}
}
\description{
This function provies a simple plug and play option for user to compare the
SFM estimates by lm and lmrobdetMM functions, using the fit.models framework.
This will allow for an easier comparison using charts and tables
}
\examples{

data(managers)
    chart.SFM(managers[,1,drop=FALSE], 
			managers[,8,drop=FALSE], 
			Rf=.035/12)	
   chart.SFM(managers[,1,drop=FALSE], 
			managers[,8,drop=FALSE], 
			Rf=managers[,10,drop=FALSE], 
			fit.models.chart=T)	
   chart.SFM(managers[,1,drop=FALSE], 
			managers[,8,drop=FALSE], 
			Rf=.035/12, fit.models.chart=T)	
   chart.SFM(managers[,1,drop=FALSE], 
			managers[,8,drop=FALSE], Rf=.035/12, 
			fit.models.chart=T, which.plots=c(1,5,6))
   chart.SFM(managers[,1,drop=FALSE], 
			managers[,8,drop=FALSE], 
			Rf=managers[,10,drop=FALSE], 
			fit.models.chart=T, which.plots=c(1,5,6))	
   chart.SFM(managers[,1,drop=FALSE], 
			managers[,8,drop=FALSE], 
			Rf=.035/12, legend.loc="bottomright")	
   chart.SFM(managers[,1,drop=FALSE], 
			managers[,8,drop=FALSE], 
			Rf=.035/12, main = "OLS vs MM estimator",
			legend.loc="bottomleft")	
   chart.SFM(managers[,1,drop=FALSE], 
			managers[,8,drop=FALSE], 
			Rf=.035/12, main = "OLS vs MM estimator",
			ylim=c(-1,1), xlim=c(-0.5,0.5),
			legend.loc="bottomleft")	
   chart.SFM(managers[,1,drop=FALSE], 
			managers[,8,drop=FALSE], 
			Rf=.035/12, main = "OLS vs MM estimator",
			ylim=c(-1,1), xlim=c(-0.5,0.5),
			family="opt", legend.loc="bottomleft")	
   chart.SFM(managers[,1,drop=FALSE], 
			managers[,8,drop=FALSE], 
			Rf=.035/12, makePct=T, 
			family = "b", lm.outliers=T)	
   
}
\references{
Sharpe, W.F. Capital Asset Prices: A theory of market
equilibrium under conditions of risk. \emph{Journal of finance}, vol 19,
1964, 425-442. \cr Ruppert, David. \emph{Statistics and Finance, an
Introduction}. Springer. 2004. \cr Bacon, Carl. \emph{Practical portfolio
performance measurement and attribution}. Wiley. 2004. \cr
}
\author{
Doug Martin, Dan Xia, Dhairya Jain
}
