% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSeedValue.R
\name{checkSeedValue}
\alias{checkSeedValue}
\title{Check 'seedValue' to ensure it is compatible with coredata_content attribute of 'R' (an xts object)}
\usage{
checkSeedValue(R, seedValue)
}
\arguments{
\item{R}{an xts object}

\item{seedValue}{a numeric scalar indicating the (usually initial) index level or price of the series}
}
\description{
Check 'seedValue' to ensure it is compatible with coredata_content attribute of 'R' (an xts object)
}
\author{
Erol Biceroglu
}
