% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAPM.coefficients.R
\name{.coefficients}
\alias{.coefficients}
\title{Wrapper for SFM's regression models.}
\usage{
.coefficients(xRa, xRb, subset, ..., method = "LS", family = "mopt")
}
\arguments{
\item{xRa}{an xts, vector, matrix, data frame, timeSeries or zoo object of
excess asset returns}

\item{xRb}{excess return vector of the benchmark asset}

\item{subset}{a logical vector}

\item{...}{}

\item{method}{Which linear model to use for SFM regression}

\item{family}{If method is "Rob", then this is a string specifying the 
name of the family of loss function to be used (current valid options are 
"bisquare", "opt" and "mopt"). Incomplete entries will be matched to the 
current valid options. Defaults to "mopt".}
}
\description{
This is intended to be called using CAPM.coefficients function, and
not directly.
}
\author{
Dhairya Jain
}
