% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawdownPeak.R
\name{DrawdownPeak}
\alias{DrawdownPeak}
\title{Drawdawn peak of the return distribution}
\usage{
DrawdownPeak(R, ...)
}
\arguments{
\item{R}{an xts, vector, matrix, data frame, timeSeries or zoo object of
asset returns (Note: expects decimal returns, e.g. 0.05 for 5\%, not percentage units).}

\item{\dots}{any other passthru parameters}
}
\description{
Drawdawn peak is for each return its drawdown since the previous peak
}
\author{
Matthieu Lestel
}
