% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_ES.R
\name{ES.hist}
\alias{ES.hist}
\title{Compute sample Expected Shortfall}
\usage{
ES.hist(data, alpha = 0.1)
}
\arguments{
\item{data}{Vector of data}

\item{alpha}{Tail Probability}
}
\value{
sample ES
}
\description{
Compute sample Expected Shortfall
}
\examples{
ES.hist(rnorm(10))
}

