% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfluenceFunctions.R
\name{ES.IF}
\alias{ES.IF}
\title{Influence Function of Expected Shortfall (ES)}
\usage{
ES.IF(data, ..., alpha = 0.1)
}
\arguments{
\item{data}{Vector of the data}

\item{...}{other parameters}

\item{alpha}{Tail Probability}
}
\value{
IF of ES
}
\description{
Influence Function of Expected Shortfall (ES)
}
\examples{
ES.IF(rnorm(10))
}
\author{
Xin Chen, \email{chenx26@uw.edu}
}

