% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimatorES.R
\name{EstimatorSE}
\alias{EstimatorSE}
\title{Wrapper function for computing the standard error of risk/performance measure estimators}
\usage{
EstimatorSE(data, ..., estimator.fun = c("Mean", "SD", "VaR", "ES", "SR",
  "SoR", "STARR"), se.method = c("none", "IFiid", "IFcor", "BOOTiid",
  "BOOTcor"))
}
\arguments{
\item{data}{xts object of the data}

\item{...}{extra arugements to be passed to lower level functions, see details}

\item{estimator.fun}{a character string indicating which risk/performancem measure's standard error
should be computed. One of \code{"Mean"} (default), \code{"SD"}, \code{"VaR"}, \code{"ES"},
\code{"SR"}, \code{"SoR"}, \code{"STARR"}.}

\item{se.method}{a character string indicating which method should be used to compute
the standard error of the estimated standard deviation. One of \code{"none"} (default),
\code{"IFiid"}, \code{"IFcor"}, \code{"BOOTiid"}, \code{"BOOTcor"}. Currely, only \code{"IFiid"}
is implemented.}
}
\value{
the standard error of the specified risk/performance measure using the specified method
}
\description{
Wrapper function for computing the standard error of risk/performance measure estimators
}
\examples{
data(edhec)
EstimatorSE(edhec,estimator.fun="SD",se.method="IFiid")
}
\author{
Xin Chen
}

