% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_EN.R
\name{myperiodogram}
\alias{myperiodogram}
\title{Compute the periodogram as defined in H&W 1981}
\usage{
myperiodogram(data, max.freq = 0.5)
}
\arguments{
\item{data}{Vector of data}

\item{max.freq}{Maximum frequency to be computed}
}
\value{
list of frequencies and corresponding periodograms
}
\description{
Compute the periodogram as defined in H&W 1981
}
\examples{
myperiodogram(rnorm(10))
}
\author{
Xin Chen
}

