% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MM.NCE.R
\name{NCE}
\alias{NCE}
\alias{MM.NCE}
\title{Functions for calculating the nearest comoment estimator for financial time series}
\usage{
MM.NCE(R, as.mat = TRUE, ...)
}
\arguments{
\item{R}{an xts, vector, matrix, data frame, timeSeries or zoo object of
asset returns (with mean zero)}

\item{as.mat}{TRUE/FALSE whether to return the full moment matrix or only
the vector with the unique elements (the latter is advised for speed), default
TRUE}

\item{\dots}{any other passthru parameters}
}
\description{
calculates NCE covariance, coskewness and cokurtosis matrices
}
\details{
The coskewness and cokurtosis matrices are defined as the matrices of dimension 
p x p^2 and p x p^3 containing the third and fourth order central moments. They
are useful for measuring nonlinear dependence between different assets of the 
portfolio and computing modified VaR and modified ES of a portfolio.

The nearest comoment estimator is a way to estimate the covariance, coskewness and
cokurtosis matrix by means of a latent multi-factor model. The method is proposed in 
CITE TODO.
}
\examples{

data(edhec)

# NCE estimation
TODO

}
\references{
TODO
}
\seealso{
\code{\link{CoMoments}} \cr \code{\link{ShrinkageMoments}} \cr \code{\link{StructuredMoments}}  
\cr \code{\link{EWMAMoments}} \cr \code{\link{MCA}}
}
\author{
Dries Cornilly
}
\concept{
co-moments

moments
}
