\name{PerformanceAnalytics-package}
\alias{PerformanceAnalytics-package}
\alias{PerformanceAnalytics}
\docType{package}
\title{
Econometric tools for performance and risk analysis.
}
\description{
Library of econometric functions for performance and risk analysis. This library aims to aid practitioners and researchers in utilizing the latest research in analysis of non-normal return streams.  In general, this library is most tested on return (rather than price) data on a monthly scale, but most functions will work with daily or irregular return data as well.
}
\details{
\tabular{ll}{
Package: \tab PerformanceAnalytics\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2007-02-01\cr
License: \tab GPL\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Peter Carl
Brian G. Peterson

Maintainer: Brian G. Peterson <brian@braverock.com>
}

\references{
Lhabitant F. \emph{Hedge Funds: Quantitative Insights}. Wiley 2004.
Tsay, R. \emph{anaysis of Financial Time Series}. Wiley 2001.
Zivot, E. and Wang, Z. \emph{Modeling Financial Time Series with S-Plus: second edition}. Springer. 2006.
}

\keyword{ package }
\seealso{}
\examples{
}
