\name{prices}
\docType{data}
\alias{prices}
\title{Selected Price Series Example Data}
\description{
A object returned by get.hist.quote of price data for use in the example for \code{\link{Return.calculate}}
}
\usage{prices}
\format{\R variable 'prices'}
\examples{
data(prices)

#preview the data
head(prices)

}
\keyword{datasets}
