% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measures.R
\name{RES}
\alias{RES}
\title{Compute sample Robust Expected Shortfall}
\usage{
RES(data, alpha = 0.01, beta = 0.1)
}
\arguments{
\item{data}{Vector of data}

\item{alpha}{Lower Tail Probability}

\item{beta}{Upper Tail Probability}
}
\value{
sample RES
}
\description{
Compute sample Robust Expected Shortfall
}
\examples{
RES(rnorm(1000))
}
\author{
Xin Chen, \email{chenx26@uw.edu}
}

