% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfluenceFunctions.R
\name{SD.IF}
\alias{SD.IF}
\title{Influence Function of Standard Deviation}
\usage{
SD.IF(data)
}
\arguments{
\item{data}{Vector of the data}
}
\value{
IF of SD
}
\description{
Influence Function of Standard Deviation
}
\examples{
SD.IF(rnorm(10))
}

