% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimatorES.R
\name{SE.BOOT.cor}
\alias{SE.BOOT.cor}
\title{Compute the standard error of the measure by tsboot()}
\usage{
SE.BOOT.cor(x, myfun, myfun.IF, ..., nsim = 100, sim = "fixed",
  l = round(length(x)/5))
}
\arguments{
\item{x}{vector of data}

\item{myfun}{measure}

\item{myfun.IF}{not used}

\item{...}{other parameters}

\item{nsim}{number of replicates}

\item{sim}{the type of simulation}

\item{l}{the length of the fixed block}
}
\value{
standard error
}
\description{
Compute the standard error of the measure by tsboot()
}
\examples{
SE.BOOT.cor(x = rnorm(100), myfun = mean, nsim = 100)
}
\author{
Xin Chen, \email{chenx26@uw.edu}
}

