% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimatorES.R
\name{SE.BOOT.iid}
\alias{SE.BOOT.iid}
\title{Compute the standard error of the measure by iid bootstrapping}
\usage{
SE.BOOT.iid(x, myfun, myfun.IF, ..., nsim = 100)
}
\arguments{
\item{x}{vector of data}

\item{myfun}{measure}

\item{myfun.IF}{not used}

\item{...}{other parameters}

\item{nsim}{number of replicates}
}
\value{
standard error
}
\description{
Compute the standard error of the measure by iid bootstrapping
}
\examples{
SE.BOOT.iid(x = rnorm(100), myfun = mean, nsim = 100)
}
\author{
Xin Chen, \email{chenx26@uw.edu}
}

