% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_ES.R
\name{SE.ES.boot.cor}
\alias{SE.ES.boot.cor}
\title{Compute standard error of ES via bootstrapping for data (possibily serially correlated)}
\usage{
SE.ES.boot.cor(data, ..., alpha = 0.05, segment.length = length(data)/5)
}
\arguments{
\item{data}{vector of data}

\item{...}{parameters passed from upper calls}

\item{alpha}{tail probability}

\item{segment.length}{the length of the fixed block for bootstrapping}
}
\value{
SE of ES
}
\description{
Compute standard error of ES via bootstrapping for data (possibily serially correlated)
}
\examples{
SE.ES.boot.cor(rnorm(100), alpha=0.1)
}

