% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_ES.R
\name{SE.ES.boot.cor.xts}
\alias{SE.ES.boot.cor.xts}
\title{Wrapper function to compute the standard error(s) of the ES for an xts object
using time series bootstrapping}
\usage{
SE.ES.boot.cor.xts(x, ..., alpha = 0.05)
}
\arguments{
\item{x}{the xts object}

\item{...}{other parameters}

\item{alpha}{tail probability}
}
\value{
standard error(s) of the xts object
}
\description{
The operation is performed column wise.
}
\examples{
data(edhec)
SE.ES.boot.cor.xts(edhec)
}

