% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_ES.R
\name{SE.ES.boot.iid}
\alias{SE.ES.boot.iid}
\title{Compute standard error of ES via bootstrapping for iid data}
\usage{
SE.ES.boot.iid(data, ..., alpha = 0.05)
}
\arguments{
\item{data}{vector of data}

\item{...}{parameters passed from upper calls}

\item{alpha}{tail probability}
}
\value{
SE of ES
}
\description{
Compute standard error of ES via bootstrapping for iid data
}
\examples{
SE.ES.boot.iid(rnorm(10))
}

