% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_ES.R
\name{SE.ES.cor}
\alias{SE.ES.cor}
\title{Compute the standard error of ES estimator for correlated data vector using
GLM-EN method}
\usage{
SE.ES.cor(x, d = 5, alpha.lasso = 0.5, keep = 1, alpha = 0.05)
}
\arguments{
\item{x}{the vector for the data}

\item{d}{the maximum order of the polynomial}

\item{alpha.lasso}{the weight for lasso vs elastic net}

\item{keep}{the portion of peridograms to use to fit the polynomial}

\item{alpha}{the tail probability of ES estimator}
}
\value{
the SE of the ES estimtor for the data
}
\description{
Compute the standard error of ES estimator for correlated data vector using
GLM-EN method
}

