% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_ES.R
\name{SE.ES.cor.xts}
\alias{SE.ES.cor.xts}
\title{Wrapper function to compute the standard error(s) of the ES for an xts object
using GLM-EN method}
\usage{
SE.ES.cor.xts(x, alpha = 0.05)
}
\arguments{
\item{x}{the xts object}

\item{alpha}{tail probability}
}
\value{
standard error(s) of the xts object
}
\description{
The operation is performed column wise.
}

