% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_ES.R
\name{SE.ES.iid}
\alias{SE.ES.iid}
\title{Compute standard error of expected shortfall}
\usage{
SE.ES.iid(data, alpha = 0.05)
}
\arguments{
\item{data}{vector of data}

\item{alpha}{tail probability}
}
\value{
SE of ES
}
\description{
Compute standard error of expected shortfall
}
\examples{
SE.ES.iid(rnorm(10))
}

