% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_EN.R
\name{SE.GLM.LASSO}
\alias{SE.GLM.LASSO}
\title{Compute the variance of sample mean of the data}
\usage{
SE.GLM.LASSO(data, d = 5, alpha = 1, keep = 1)
}
\arguments{
\item{data}{Vector of data}

\item{d}{Maximum order of the polynomial}

\item{alpha}{Weight of regularization. alpha=1 LASSO, alpha=0 Ridge}

\item{keep}{Percentage of periodograms to use for regression}
}
\value{
variance of the sample mean. To get the asymptotic variance, multiply by length(data)
}
\description{
Using GLM with L1 regularization to fit a polynomial from the periodogram then the intercept of the fitted polynomial is the spectral density p(0) at frequency 0
}
\examples{
require(h2o)
h2o.init()
SE.GLM.LASSO(rnorm(10))
# h2o.shutdown(prompt=FALSE)
}

