% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimatorES.R
\name{SE.IF.cor}
\alias{SE.IF.cor}
\title{Compute the standard error using GLM-EN approach for serially correlated data}
\usage{
SE.IF.cor(x, myfun.IF, ..., d = 5, alpha.lasso = 0.5, keep = 1)
}
\arguments{
\item{x}{the vector of data}

\item{myfun.IF}{the influene function of the measure}

\item{...}{other parameters}

\item{d}{maximum order of the polynomial}

\item{alpha.lasso}{weight for the elastic net}

\item{keep}{portion of frequencies to be used}
}
\value{
the standard error of the measure
}
\description{
Compute the standard error using GLM-EN approach for serially correlated data
}
\author{
Xin Chen, \email{chenx26@uw.edu}
}

