% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimatorES.R
\name{SE.IF.iid}
\alias{SE.IF.iid}
\title{Compute the standard error using influence function approach for a vector}
\usage{
SE.IF.iid(x, myfun.IF, ...)
}
\arguments{
\item{x}{vector of data}

\item{myfun.IF}{influence function of the measure}

\item{...}{other parameters used to compute the influence function}
}
\value{
standard error of the measure
}
\description{
Compute the standard error using influence function approach for a vector
}
\examples{
SE.IF.iid(rnorm(100), SD.IF)
}
\author{
Xin Chen, \email{chenx26@uw.edu}
}

