% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimatorES.R
\name{SE.xts}
\alias{SE.xts}
\title{compute the standard error for the xts object}
\usage{
SE.xts(x, se.fun, myfun, myfun.IF, ...)
}
\arguments{
\item{x}{the xts object of the data}

\item{se.fun}{the function used to compute the standard error}

\item{myfun}{the measure}

\item{myfun.IF}{the influence function of the measure}

\item{...}{other parameters}
}
\value{
standard errors
}
\description{
compute the standard error for the xts object
}
\examples{
data(edhec)
SE.xts(edhec, SE.IF.iid, sd, SD.IF)
}
\author{
Xin Chen, \email{chenx26@uw.edu}
}

