% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAPM.coefficients.R
\name{SFM.coefficients}
\alias{SFM.coefficients}
\alias{CAPM.coefficients}
\title{Calculate single factor model (CAPM) alpha and beta coefficients}
\usage{
SFM.coefficients(Ra, Rb, subset = TRUE, Rf = 0, ...)
}
\arguments{
\item{Ra}{an xts, vector, matrix, data frame, timeSeries or zoo object of
asset returns}

\item{Rb}{return vector of the benchmark asset}

\item{subset}{a logical vector representing the set of observations to be used in regression}

\item{Rf}{risk free rate, in same period as your returns}

\item{\dots}{Parameters like method, family and other parameters like max.it or bb 
for lmrobdetMM regression.}

\item{method}{(Optional): string representing linear regression model, "LS" for Least Squares
and "Rob" for robust}

\item{family}{(Optional): 
If method == "Rob": 
  This is a string specifying the name of the family of loss function
  to be used (current valid options are "bisquare", "opt" and "mopt").
  Incomplete entries will be matched to the current valid options. 
  Defaults to "mopt".
Else: the parameter is ignored}
}
\description{
The single factor model or CAPM is the beta of an asset to the variance 
and covariance of an initial portfolio. Used to determine diversification potential.
"Alpha" purports to be a measure of a manager's skill by measuring the
portion of the managers returns that are not attributable to "Beta", or the
portion of performance attributable to a benchmark.
}
\details{
This function is designed to be used a a wrapper for SFM's regression models. 
Using this, one can easily fit different types of linear models 
like Ordinary Least Squares or Robust Estimators like Huber.

\deqn{\beta_{a,b}=\frac{CoV_{a,b}}{\sigma_{b}}=\frac{\sum((R_{a}-\bar{R_{a}})(R_{b}-\bar{R_{b}}))}{\sum(R_{b}-\bar{R_{b}})^{2}}}{beta
= cov(Ra,Rb)/var(R)}

Ruppert(2004) reports that this equation will give the estimated slope of
the linear regression of \eqn{R_{a}}{Ra} on \eqn{R_{b}}{Rb} and that this
slope can be used to determine the risk premium or excess expected return
(see Eq. 7.9 and 7.10, p. 230-231).
}
\examples{

data(managers)
     SFM.coefficients(managers[,1], 
			     managers[,8], method="Rob", 
			     family="mopt")
     SFM.coefficients(managers[,1], 
			     managers[,8:9], method="Both") 
     SFM.coefficients(managers[,1:6], 
			     managers[,8:9], 
			     Rf = managers[,10])
			SFM.coefficients(managers[,1:6], 
			     managers[,8:9],
			     Rf = managers[,10],
			     method="Both")
			SFM.coefficients(managers[,1:6], 
			     managers[,8:9], 
			     Rf=.035/12, method="Rob", 
			     family="opt", bb=0.25, 
			     max.it=200)
     SFM.coefficients(managers[, "HAM2"], 
          managers[, "SP500 TR"], 
			     Rf = managers[,10],
			     method="Rob", family="mopt")
			     
     sfmRob <- SFM.coefficients(managers[, "HAM2"], 
          managers[, "SP500 TR"], 
			     Rf=.035/12, method="Rob", family="opt", 
			     bb=0.25, max.it=200)
			     sfmRob$beta
			     sfmRob$intercept
			     sfmRob$model
			     
     sfmLSRob <- SFM.coefficients(managers[, "HAM2"], 
          managers[, "SP500 TR"], 
          Rf = managers[, "US 3m TR"], 
          method="Both", family="opt") 
          sfmLSRob$robust
          sfmLSRob$ordinary
          sfmLSRob$robust$beta
          sfmLSRob$robust$intercept
          sfmLSRob$robust$model
          sfmLSRob$ordinary$beta
          sfmLSRob$ordinary$intercept
          sfmLSRob$ordinary$model
          
		  table <- SFM.coefficients(managers[,1:6], 
			     managers[,8:9],
			     Rf = managers[,10],
			     method="LS")
		       table["Intercept, Beta, Model: SP500 TR","HAM1"]
		       table["Intercept, Beta, Model: SP500 TR","HAM1"][[1]]
		       table["Intercept, Beta, Model: SP500 TR","HAM1"][[1]]$beta
 	       table["Intercept, Beta, Model: SP500 TR","HAM1"][[1]]$intercept
	         table["Intercept, Beta, Model: SP500 TR","HAM1"][[1]]$model
	         table["Intercept, Beta, Model: SP500 TR","HAM1"][[1]]$model$residuals
  	  
  	  table <- SFM.coefficients(managers[,1:6], 
			     managers[,8:9], Rf = managers[,10],
			     method="Rob", family="opt")
		       table["Intercept, Beta, Model: SP500 TR","HAM1"]
		       table["Intercept, Beta, Model: SP500 TR","HAM1"][[1]]
		       table["Intercept, Beta, Model: SP500 TR","HAM1"][[1]]$beta
 	       table["Intercept, Beta, Model: SP500 TR","HAM1"][[1]]$intercept
	         table["Intercept, Beta, Model: SP500 TR","HAM1"][[1]]$model
	         table["Intercept, Beta, Model: SP500 TR","HAM1"][[1]]$model$residuals
  	  
  	  table <- SFM.coefficients(managers[,1:6], 
			     managers[,8:9],
			     Rf = managers[,10],
			     method="Both")
		       table["Intercept, Beta, Model: SP500 TR","HAM1"]
		       table["Intercept, Beta, Model: SP500 TR","HAM1"][[1]]
		       table["Intercept, Beta, Model: SP500 TR","HAM1"][[1]]$ordinary$beta
 	       table["Intercept, Beta, Model: SP500 TR","HAM1"][[1]]$ordinary$intercept
	         table["Intercept, Beta, Model: SP500 TR","HAM1"][[1]]$ordinary$model
		       table["Intercept, Beta, Model: SP500 TR","HAM1"][[1]]$robust$beta
		       table["Intercept, Beta, Model: SP500 TR","HAM1"][[1]]$robust$intercept
		       table["Intercept, Beta, Model: SP500 TR","HAM1"][[1]]$robust$model
	         table["Intercept, Beta, Model: SP500 TR","HAM1"][[1]]$ordinary
	         table["Intercept, Beta, Model: SP500 TR","HAM1"][[1]]$robust
		       table["Intercept, Beta, Model: US 10Y TR","HAM2"]
		       table["Intercept, Beta, Model: US 10Y TR","HAM2"][[1]]
		       table["Intercept, Beta, Model: US 10Y TR","HAM2"][[1]]$ordinary$beta
 	       table["Intercept, Beta, Model: US 10Y TR","HAM2"][[1]]$ordinary$intercept
	         table["Intercept, Beta, Model: US 10Y TR","HAM2"][[1]]$ordinary$model
		       table["Intercept, Beta, Model: US 10Y TR","HAM2"][[1]]$robust$beta
		       table["Intercept, Beta, Model: US 10Y TR","HAM2"][[1]]$robust$intercept
		       table["Intercept, Beta, Model: US 10Y TR","HAM2"][[1]]$robust$model
	         table["Intercept, Beta, Model: US 10Y TR","HAM2"][[1]]$ordinary
	         table["Intercept, Beta, Model: US 10Y TR","HAM2"][[1]]$robust
  	  
}
\references{
Sharpe, W.F. Capital Asset Prices: A theory of market
equilibrium under conditions of risk. \emph{Journal of finance}, vol 19,
1964, 425-442. \cr Ruppert, David. \emph{Statistics and Finance, an
Introduction}. Springer. 2004. \cr Bacon, Carl. \emph{Practical portfolio
performance measurement and attribution}. Wiley. 2004. \cr
}
\seealso{
\code{\link{BetaCoVariance}} \code{\link{SFM.alpha}}
\code{\link{CAPM.utils}} \code{\link{SFM.beta}}
}
\author{
Dhairya Jain
}
