% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAPM.chart.doug.R
\name{SFM.compare}
\alias{SFM.compare}
\alias{CAPM.compare}
\title{Compare CAPM estimated using robust estimators with that estimated by OLS}
\usage{
SFM.compare(
  Ra,
  Rb,
  Rf = 0,
  mainText = NULL,
  ylimits = NULL,
  family = "mopt",
  legendPos = "topleft",
  goodOutlier = F,
  makePct = FALSE
)
}
\arguments{
\item{Ra}{an xts, vector, matrix, data frame, timeSeries or zoo object of
asset returns}

\item{Rb}{return vector of the benchmark asset}

\item{Rf}{risk free rate, in same period as your returns}

\item{mainText}{MainText for the plot}

\item{ylimits}{Limits on the y-axis of the plots. Defaults to min-max}

\item{family}{(Optional): 
This is a string specifying the name of the family of loss function
to be used (current valid options are "bisquare", "opt" and "mopt").
Incomplete entries will be matched to the current valid options. 
Defaults to "mopt".}

\item{legendPos}{Position of legends. See plot() function for more info.}

\item{makePct}{If Returns should be converted to percentage. Defaults to False}
}
\description{
This function provies a simple plug and play option for user to compare the
SFM estimates by lm and lmrobdetMM functions, using the fit.models framework.
This will allow for an easier comparison using charts and tables
}
\examples{

data(managers)
    SFM.compare(managers[,1,drop=FALSE], 
			managers[,8,drop=FALSE], 
			Rf=.035/12)	
}
\references{
Sharpe, W.F. Capital Asset Prices: A theory of market
equilibrium under conditions of risk. \emph{Journal of finance}, vol 19,
1964, 425-442. \cr Ruppert, David. \emph{Statistics and Finance, an
Introduction}. Springer. 2004. \cr Bacon, Carl. \emph{Practical portfolio
performance measurement and attribution}. Wiley. 2004. \cr
}
\author{
Doug Martin, Dan Xia, Dhairya Jain
}
