% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAPM.fit.models.R
\name{SFM.fit.models}
\alias{SFM.fit.models}
\title{Compare CAPM estimated using robust estimators with that estimated by OLS}
\usage{
SFM.fit.models(Ra, Rb, Rf = 0)
}
\arguments{
\item{Ra}{an xts, vector, matrix, data frame, timeSeries or zoo object of
asset returns}

\item{Rb}{return vector of the benchmark asset}

\item{Rf}{risk free rate, in same period as your returns}
}
\description{
This function provies a simple plug and play option for user to compare the
SFM estimates by lm and lmrobdetMM functions, using the fit.models framework.
This will allow for an easier comparison using charts and tables
}
\examples{

# First we load the data
    data(managers)
    SFM.fit.models(managers[,1], 
			managers[,8,drop=FALSE], 
			Rf = managers[,10,drop=FALSE])
    SFM.fit.models(managers[,6], 
			managers[,8,drop=FALSE], 
			Rf=.035/12) 
		 SFM.fit.models(managers[, "HAM2", drop=FALSE], 
			managers[, "SP500 TR", drop=FALSE], 
			Rf = managers[, "US 3m TR", drop=FALSE])
		SFM.fit.models(managers[, "HAM2", drop=FALSE], 
			managers[, "SP500 TR", drop=FALSE], 
			Rf = 0.035/12)
		SFM.fit.models(managers[, "HAM2", drop=FALSE], 
			managers[, "SP500 TR", drop=FALSE], 
			Rf = managers[,10,drop=FALSE])
		
}
\author{
Dhairya Jain
}
