% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfluenceFunctions.R
\name{SoR.const.IF}
\alias{SoR.const.IF}
\title{Influence Function of Sortino Ratio with Constant Threshold}
\usage{
SoR.const.IF(data, ..., const = 0)
}
\arguments{
\item{data}{Vector of data}

\item{...}{other parameters}

\item{const}{The threshodl}
}
\value{
IF of SoR
}
\description{
Influence Function of Sortino Ratio with Constant Threshold
}
\examples{
SoR.const.IF(rnorm(10),MAR = 0.1)
}
\author{
Xin Chen, \email{chenx26@uw.edu}
}

