% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measures.R
\name{SoR.const}
\alias{SoR.const}
\title{Compute sample Sortino Ratio with constant threshold}
\usage{
SoR.const(data, ..., MAR = 0)
}
\arguments{
\item{data}{Vector of data}

\item{...}{other parameters}

\item{MAR}{Constant Threshold}
}
\value{
SoR
}
\description{
Compute sample Sortino Ratio with constant threshold
}
\examples{
SoR.const.IF(rnorm(10),MAR = 0.1)
}
\author{
Xin Chen, \email{chenx26@uw.edu}
}

