% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfluenceFunctions.R
\name{SoR.IF}
\alias{SoR.IF}
\title{Influence Function of Sortino Ratio with Mean Return Threshold}
\usage{
SoR.IF(data, rf = 0)
}
\arguments{
\item{data}{Vector of data}

\item{rf}{Risk-Free interest rate}
}
\value{
IF of SoR
}
\description{
Influence Function of Sortino Ratio with Mean Return Threshold
}
\examples{
SoR.IF(rnorm(10))
}

