% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfluenceFunctions.R
\name{SR.IF}
\alias{SR.IF}
\title{Compute the Influence Function value of Sharpe Ratio}
\usage{
SR.IF(data, rf = 0)
}
\arguments{
\item{data}{vector of data}

\item{rf}{risk free interest rate}
}
\value{
vector of influence function
}
\description{
Compute the Influence Function value of Sharpe Ratio
}
\examples{
SR.IF(rnorm(10))
}

