% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measures.R
\name{SR}
\alias{SR}
\title{Compute sample Sharpe Ratio from data}
\usage{
SR(data, rf = 0)
}
\arguments{
\item{data}{Vector of data}

\item{rf}{Risk-free rate}
}
\value{
Value of Sample Sharpe Ratio
}
\description{
Compute sample Sharpe Ratio from data
}
\examples{
SR(rnorm(10))
}
\author{
Xin Chen, \email{chenx26@uw.edu}
}

