% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfluenceFunctions.R
\name{STARR.IF}
\alias{STARR.IF}
\title{Influence Function of STARR Ratio}
\usage{
STARR.IF(data, ..., alpha = 0.1, rf = 0)
}
\arguments{
\item{data}{Vector of data}

\item{...}{other parameters}

\item{alpha}{Tail Probability}

\item{rf}{risk free interest rate}
}
\value{
IF of STARR
}
\description{
Influence Function of STARR Ratio
}
\examples{
STARR.IF(rnorm(10))
}
\author{
Xin Chen, \email{chenx26@uw.edu}
}

