% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measures.R
\name{STARR}
\alias{STARR}
\title{Compute sample STARR}
\usage{
STARR(data, alpha = 1, rf = 0)
}
\arguments{
\item{data}{vector of data}

\item{alpha}{tail probability}

\item{rf}{risk-free interest rate}
}
\value{
sample STARR
}
\description{
Compute sample STARR
}
\examples{
STARR(rnorm(10))
}
\author{
Xin Chen, \email{chenx26@uw.edu}
}

