% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STARR_SE.R
\name{STARR.SE}
\alias{STARR.SE}
\title{Wrapper function that computes STARR and the standard error of the estimate}
\usage{
STARR.SE(data, ..., alpha = 0.1, rf = 0, se.method = "none")
}
\arguments{
\item{data}{data}

\item{alpha}{tail probability}

\item{rf}{risk free rate}

\item{se.method}{a character string indicating which method should be used to compute
the standard error of the estimated standard deviation. One of \code{"none"} (default),
\code{"IFiid"}, \code{"IFcor"}, \code{"BOOTiid"}, \code{"BOOTcor"}. Currently, it works
only when \code{method="historical"} and \code{portfolio_method="single"}.}

\item{\dots}{any other passthru parameters. This include two types of parameters.
The first type is parameters associated with the risk/performance measure, such as tail
probability for VaR and ES. The second type is the parameters associated with the metohd
used to compute the standard error. See \code{\link{SE.IF.iid}}, \code{\link{SE.IF.cor}},
\code{\link{SE.BOOT.iid}}, \code{\link{SE.BOOT.cor}} for details.}
}
\value{
a vector or a list depending on se.method
}
\description{
Wrapper function that computes STARR and the standard error of the estimate
}
\author{
Xin Chen, \email{chenx26@uw.edu}
}

