% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_weights.R
\docType{data}
\name{test_weights}
\alias{test_weights}
\title{Sample sector weights for use by unit tests}
\format{
A data frame with 5 rows and 10 variables:
\describe{
  \item{Sector1}{weight in Sector1, numeric}
  \item{Sector2}{weight in Sector2, numeric}
  \item{Sector3}{weight in Sector3, numeric}
  \item{Sector4}{weight in Sector4, numeric}
  \item{Sector5}{weight in Sector5, numeric}
  \item{Sector6}{weight in Sector6, numeric}
  \item{Sector7}{weight in Sector7, numeric}
  \item{Sector8}{weight in Sector8, numeric}
  \item{Sector9}{weight in Sector9, numeric}
  \item{Sector10}{weight in Sector10, numeric}
}
}
\usage{
test_weights
}
\description{
A dataset containing weights for 10 sectors over 5 month-end periods.
}
\keyword{datasets}
