% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR.backtest.R
\name{VaR.backtest}
\alias{VaR.backtest}
\title{VaR Backtest}
\usage{
VaR.backtest(R, VaR, p = 0.95)
}
\arguments{
\item{R}{an xts, vector, matrix, data frame, timeSeries or zoo object of
actual asset returns}

\item{VaR}{estimated Value at Risk scalar or vector corresponding to the same periods}

\item{p}{confidence level for calculation, default p=.95}
}
\value{
A list containing:
\itemize{
  \item \code{expected_exceedances} the expected number of VaR breaches.
  \item \code{actual_exceedances} the number of actual VaR breaches.
  \item \code{p.value} the p-value from the binomial proportion test.
}
}
\description{
Provides a simple binomial backtest for a Value at Risk (VaR) model.

This function performs a Kupiec Proportion of Failures (POF) test to evaluate
whether the number of times returns breached the VaR threshold matches the expected
frequency \code{1-p}.

The original implementation was provided in the archived \kbd{VaR} package by Talgat Daniyarov.
}
\examples{

data(edhec)
# calculate VaR at 95 percent confidence
v <- as.numeric(VaR(edhec[, 1], p = 0.95, method = "historical"))

# evaluate the backtest
VaR.backtest(edhec[, 1], v, p = 0.95)

}
\references{
Daniyarov, T. \emph{VaR: Value at Risk estimation}. CRAN Archive. 2004.
\url{https://cran.r-project.org/src/contrib/Archive/VaR/}
}
\seealso{
\code{\link{VaR}} \cr
}
\author{
Brian G. Peterson, Talgat Daniyarov
}
