% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfluenceFunctions.R
\name{VaR.IF}
\alias{VaR.IF}
\title{Compute the influence function of value-at-risk}
\usage{
VaR.IF(data, ..., alpha = 0.1)
}
\arguments{
\item{data}{The vector of data}

\item{...}{other parameters}

\item{alpha}{The tail probability}
}
\value{
Influence Function of VaR
}
\description{
Compute the influence function of value-at-risk
}
\examples{
VaR.IF(rnorm(10))
}
\author{
Xin Chen, \email{chenx26@uw.edu}
}

