% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{xtsboot}
\alias{xtsboot}
\title{Uses sample function to construct bootstrap error in xts objects
The metric is computed for rep times sizefrac*(size) sized samples and the variance of the metric over these tries is returned
This is NOT filtered for NA values outside of filtering in the functions we are passing to. This will be fixed.
Can be used as a generalized wrapper for metrics that return numeric values from xts}
\usage{
xtsboot(x, fun = NULL, sizefrac = 0.5, rep = 100, flag = 0)
}
\arguments{
\item{x}{the xts object ( note : no benchmark xts implemented yet ) , fun = the function, such as Sharpe Ratio}

\item{sizefrac}{= the size of each bootstrap sample as a fraction of total size, rep = repetitions, flag - 0,1,2 for var,mean,median resp.}
}
\description{
Uses sample function to construct bootstrap error in xts objects
The metric is computed for rep times sizefrac*(size) sized samples and the variance of the metric over these tries is returned
This is NOT filtered for NA values outside of filtering in the functions we are passing to. This will be fixed.
Can be used as a generalized wrapper for metrics that return numeric values from xts
}

