% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zerofill.R
\name{zerofill}
\alias{zerofill}
\title{zerofill}
\usage{
zerofill(x)
}
\arguments{
\item{x}{time series to zero fill}
}
\description{
Fill NA's with zeros in a time series to allow analysis when the data must
be complete.
}
\details{
Note that this function has risks, use carefully.  Complete data is
preferred.  Barring that, filling a small percentage of results in the
middle of a large set is unlikely to cause problems. Barring that, realize
that this will skew your results.
}

