% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DrawdownPeak}
\alias{DrawdownPeak}
\title{Drawdawn peak of the return distribution}
\usage{
DrawdownPeak(R, ...)
}
\arguments{
\item{R}{an xts, vector, matrix, data frame, timeSeries or zoo object of
asset returns}

\item{\dots}{any other passthru parameters}
}
\description{
Drawdawn peak is for each return its drawdown since the previous peak
}
\author{
Matthieu Lestel
}

