\name{edhec}
\docType{data}
\alias{edhec}
\title{EDHEC-Risk Hedge Fund Style Indices}
\description{
EDHEC composite hedge fund style index returns.
}
\usage{edhec}
\format{CSV conformed into a data.frame with monthly observations}
\source{\url{http://www.edhec-risk.com/indexes/pure_style}}
\references{
About EDHEC Alternative Indexes. December 16, 2003. EDHEC-Risk.
\url{http://www.edhec-risk.com/indexes/pure_style/about}

Vaissié, Mathieu. A Detailed Analysis of the Construction Methods and Management Principles of Hedge Fund Indices. October 2003. EDHEC.
\url{http://www.edhec-risk.com/site_edhecrisk/public/indexes/EDHEC_Publications/RISKReview1072705188065793513}
}
\keyword{datasets}
\keyword{ ts }
