% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfluenceFunctions.R
\name{mu.IF}
\alias{mu.IF}
\title{Influence Function of Mean}
\usage{
mu.IF(data)
}
\arguments{
\item{data}{Vector of the data}
}
\value{
IF of Mean
}
\description{
Influence Function of Mean
}
\examples{
mu.IF(rnorm(10))
}
\author{
Xin Chen
}

