\name{PerformanceAnalytics-package}
\alias{PerformanceAnalytics-package}
\alias{PerformanceAnalytics}
\docType{package}
\title{
Econometric tools for performance and risk analysis.
}
\description{
Library of econometric functions for performance and risk analysis of financial portfolios. This library aims to aid practitioners and researchers in using the latest research in analysis of both normal and non-normal return streams.  In general, this library is most tested on return (rather than price) data on a monthly scale, but most functions will work with daily or irregular return data as well.
}
\details{
\tabular{ll}{
Package: \tab PerformanceAnalytics\cr
Type: \tab Package\cr
Version: \tab 0.9.1\cr
Date: \tab 2007-02-26\cr
License: \tab GPL\cr
}
We created this library to include functionality that has been appearing in the academic literature on performance analysis and risk over the past several years, but had no functional equivalent in R.  In doing so, we also found it valuable to have wrapper functions for functionality easily replicated in R, so that we could access that functionality using a function with defaults and naming consistent with common usage in the finance literature.

Performance Analysis

Risk Analysis

Summary Tabular Data

Charts and Graphs

Wrapper and Utility Functions

}
\author{
Peter Carl \cr
Brian G. Peterson \cr

Maintainer: Brian G. Peterson <brian@braverock.com> \url{http://braverock.com/brian/}
}

\references{
Lhabitant F. \emph{Hedge Funds: Quantitative Insights}. Wiley. 2004. \cr
Ruppert, David. \emph{Statistics and Finance, an Introduction}. Springer. 2004. \cr
Tsay, R. \emph{Analysis of Financial Time Series}. Wiley. 2001. \cr
Zivot, E. and Wang, Z. \emph{Modeling Financial Time Series with S-Plus: second edition}. Springer. 2006. \cr
}

\keyword{ package }
\seealso{}
\examples{
}
