\name{CalculateReturns}
\alias{CalculateReturns}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ calculate simple or compound returns from prices }
\description{
calculate simple or compound returns from prices
}
\usage{
CalculateReturns(Prices, method=c("simple","compound"))
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Prices}{vector or matrix of prices}
  \item{method}{calculate "simple" or "compound" returns}
}
\value{
vector or matrix of simple or compound returns
}
\details{
}
\references{ }
\author{ Peter Carl }
\note{
}
\seealso{ }
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ts }
\keyword{ multivariate }
\keyword{ distribution }
\keyword{ models }