% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SharpeRatio.annualized}
\alias{SharpeRatio.annualized}
\title{calculate annualized Sharpe Ratio}
\usage{
SharpeRatio.annualized(R, Rf = 0, scale = NA, geometric = TRUE)
}
\arguments{
\item{R}{an xts, vector, matrix, data frame, timeSeries or zoo object of
asset returns}

\item{Rf}{risk free rate, in same period as your returns}

\item{scale}{number of periods in a year (daily scale = 252, monthly scale =
12, quarterly scale = 4)}

\item{geometric}{utilize geometric chaining (TRUE) or simple/arithmetic chaining (FALSE) to aggregate returns,
default TRUE}
}
\description{
The Sharpe Ratio is a risk-adjusted measure of return that uses standard
deviation to represent risk.
}
\details{
The Sharpe ratio is simply the return per unit of risk (represented by
variance).  The higher the Sharpe ratio, the better the combined performance
of "risk" and return.

This function annualizes the number based on the scale parameter.

\deqn{\frac{\sqrt[n]{prod(1+R_{a})^{scale}}-1}{\sqrt{scale}\cdot\sqrt{\sigma}}}

Using an annualized Sharpe Ratio is useful for comparison of multiple return
streams.  The annualized Sharpe ratio is computed by dividing the annualized
mean monthly excess return by the annualized monthly standard deviation of
excess return.

William Sharpe now recommends Information Ratio preferentially to the
original Sharpe Ratio.
}
\examples{
data(managers)
SharpeRatio.annualized(managers[,1,drop=FALSE], Rf=.035/12)
SharpeRatio.annualized(managers[,1,drop=FALSE], Rf = managers[,10,drop=FALSE])
SharpeRatio.annualized(managers[,1:6], Rf=.035/12)
SharpeRatio.annualized(managers[,1:6], Rf = managers[,10,drop=FALSE])
SharpeRatio.annualized(managers[,1:6], Rf = managers[,10,drop=FALSE],geometric=FALSE)
}
\author{
Peter Carl
}
\references{
Sharpe, W.F. The Sharpe Ratio,\emph{Journal of Portfolio
Management},Fall 1994, 49-58.
}
\seealso{
\code{\link{SharpeRatio}} \cr \code{\link{InformationRatio}} \cr
\code{\link{TrackingError}} \cr \code{\link{ActivePremium}} \cr
\code{\link{SortinoRatio}}
}
\keyword{distribution}
\keyword{models}
\keyword{multivariate}
\keyword{ts}

