% Generated by roxygen2 (4.0.1): do not edit by hand
\name{VaR.Marginal}
\alias{VaR.Marginal}
\title{VaR.Marginal}
\usage{
VaR.Marginal(R, p = 0.95, method = c("modified", "gaussian", "historical"),
  weights = NULL, mu = NULL, sigma = NULL, m3 = NULL, m4 = NULL,
  invert = TRUE, reasoncheck = FALSE, weightsFlag = TRUE)
}
\description{
Marginal Value at Risk
\code{VaR.Marginal} Calculates the VaR on the margin for a risk factor.
}

